+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Marlin adheres to [Semantic Versioning](http://semver.org/).
 
 ##### Bug Fixes
 * Don't allow for secondary sort field if no primary
+* Prevent navigation drawer tab drag and drop from allowing more than 4 tabs
 
 ## 1.0.0
   * Initial Release: Marlin brings NGA’s Maritime Safety Information datasets to your mobile device to help you navigate safely at sea. With Marlin, you can download the latest maritime safety reports and reference information before setting sail, then search and view the information on an interactive map, even when you have no cell service. Marlin automatically checks for updates when you have a data connection, and informs you what’s new.